requirements.txt
flask
google-cloud-pubsub
google-cloud-bigquery
transformers
torch